CommentRepository 생성

+++ src/main/java/site/metacoding/white/service/BoardService.java
@@ -1,18 +1,24 @@
 package site.metacoding.white.service;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import site.metacoding.white.domain.Board;
 import site.metacoding.white.domain.BoardRepository;
 import site.metacoding.white.domain.UserRepository;
 import site.metacoding.white.dto.BoardReqDto.BoardSaveReqDto;
+import site.metacoding.white.dto.BoardReqDto.BoardUpdateReqDto;
 import site.metacoding.white.dto.BoardRespDto.BoardDetailRespDto;
 import site.metacoding.white.dto.BoardRespDto.BoardSaveRespDto;
+import site.metacoding.white.dto.BoardRespDto.BoardUpdateRespDto;
 
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class BoardService {
@@ -35,26 +41,59 @@ public class BoardService {
 
     @Transactional(readOnly = true) // 세션 종료 안 됨
     public BoardDetailRespDto findById(Long id) {
-        Board boardPS = boardRepository.findById(id); // 오픈 인뷰가 false니까 조회후 세션 종료
+
+        /*
+         * Board boardPS = boardRepository.findById(id)
+         * .orElseThrow(() -> new RuntimeException("해당 " + id + "로 상세보기를 할 수 없습니다."));
+         */
+
+        Optional<Board> boardOP = boardRepository.findById(id); // 오픈 인뷰가 false니까 조회후 세션 종료
         // boardPS.getUser().getUsername(); // Lazy 로딩됨. (근데 Eager이면 이미 로딩되서 select 두번
-        BoardDetailRespDto boardDetailRespDto = new BoardDetailRespDto(boardPS);
-        System.out.println("서비스단에서 지연로딩 함. 왜? 여기까지는 디비커넥션이 유지되니까");
-        return boardDetailRespDto;
+        if (boardOP.isPresent()) {
+            BoardDetailRespDto boardDetailRespDto = new BoardDetailRespDto(boardOP.get());
+            return boardDetailRespDto;
+        } else {
+            throw new RuntimeException("해당 " + id + "로 상세보기를 할 수 없습니다.");
+        }
     }
 
     @Transactional
-    public void update(Long id, Board board) {
+    public BoardUpdateRespDto update(BoardUpdateReqDto boardUpdateReqDto) {
+        Long id = boardUpdateReqDto.getId();
         // 영속화된 데이터를 수정!
-        Board boardPS = boardRepository.findById(id);
-        boardPS.update(board.getTitle(), board.getContent());
+        Optional<Board> boardOP = boardRepository.findById(id);
+        if (boardOP.isPresent()) {
+            Board boardPS = boardOP.get();
+            boardPS.update(boardUpdateReqDto.getTitle(), boardUpdateReqDto.getContent());
+            return new BoardUpdateRespDto(boardPS);
+        } else {
+            throw new RuntimeException("해당 " + id + "로 수정을 할 수 없습니다.");
+        }
     } // 트랜젝션 종료시 -> 더티체킹을 함
 
-    public List<Board> findAll() {
-        return boardRepository.findAll();
+    @Transactional(readOnly = true)
+    public List<BoardDetailRespDto> findAll() {
+        List<Board> boardList = boardRepository.findAll();
+        List<BoardDetailRespDto> boardDetailList = new ArrayList<>();
+        // 1. List의 크기만큼 for문 돌리기
+        for (int i = 0; i < boardList.size(); i++) {
+            BoardDetailRespDto boardDetailRespDto = new BoardDetailRespDto(boardList.get(i));
+            boardDetailList.add(boardDetailRespDto);
+        }
+        // 2. Board -> DTO로 옮겨야 함
+
+        // 3. DTO를 List에 담기
+
+        return boardDetailList;
     }
 
     @Transactional
     public void deleteById(Long id) {
-        boardRepository.deleteById(id);
+        Optional<Board> boardOP = boardRepository.findById(id);
+        if (boardOP.isPresent()) {
+            boardRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("해당 " + id + "로 삭제를 할 수 없습니다.");
+        }
     }
 }
